refactor(search): replace hybrid search with pure semantic search
